+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * update `Requires at least` in plugin file
 * add VCS checkouts automatically, thanks @Raruto
+* add VCS icons as setting
 
 #### 2.4.1 / 2019-05-09
 * a11y updates for settings tabs